requirements.txt
amqp==5.0.6
ansible==2.9.2
asgiref==3.3.1
billiard==3.6.4.0
celery==5.0.5
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
Django==2.2.16
django-celery-beat==2.2.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-ranged-response==0.2.0
django-redis==4.12.1
django-rest-framework-mongoengine==3.4.1
django-simple-captcha==0.5.13
django-timezone-field==4.2.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
dnspython==1.16.0
drf-extensions==0.7.0
drf-yasg==1.20.0
eventlet==0.30.2
greenlet==1.1.1
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonpath==0.82
kombu==5.1.0
MarkupSafe==2.0.1
mongoengine==0.22.1
mysqlclient==2.0.3
packaging==21.0
Pillow==8.1.0
prompt-toolkit==3.0.19
psutil==5.8.0
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
pymongo==3.11.3
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
rich==10.9.0
ruamel.yaml==0.17.11
ruamel.yaml.clib==0.2.6
six==1.15.0
sqlparse==0.4.1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.6
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
